fix circular dep checks
