+++ packages/node-experimental/src/sdk/api.ts
@@ -19,65 +19,15 @@ import type {
   User,
 } from '@sentry/types';
 import { GLOBAL_OBJ, consoleSandbox, dateTimestampInSeconds } from '@sentry/utils';
-
 import { getScopesFromContext, setScopesOnContext } from '../utils/contextData';
+
 import type { ExclusiveEventHintOrCaptureContext } from '../utils/prepareEvent';
 import { parseEventHintOrCaptureContext } from '../utils/prepareEvent';
-import { getGlobalCarrier } from './globals';
-import { Scope } from './scope';
-import type { CurrentScopes, SentryCarrier } from './types';
+import type { Scope } from './scope';
+import { getClient, getCurrentScope, getGlobalScope, getIsolationScope } from './scope';
 
-/** Get the currently active client. */
-export function getClient<C extends Client>(): C {
-  const currentScope = getCurrentScope();
-  const isolationScope = getIsolationScope();
-  const globalScope = getGlobalScope();
-
-  const client = currentScope.getClient() || isolationScope.getClient() || globalScope.getClient();
-  if (client) {
-    return client as C;
-  }
-
-  // TODO otherwise ensure we use a noop client
-  return {} as C;
-}
-
-/** Get the current scope. */
-export function getCurrentScope(): Scope {
-  return getScopes().scope as Scope;
-}
-
-/**
- * Set the current scope on the execution context.
- * This should mostly only be called in Sentry.init()
- */
-export function setCurrentScope(scope: Scope): void {
-  getScopes().scope = scope;
-}
-
-/** Get the global scope. */
-export function getGlobalScope(): Scope {
-  const carrier = getGlobalCarrier();
-
-  if (!carrier.globalScope) {
-    carrier.globalScope = new Scope();
-  }
-
-  return carrier.globalScope as Scope;
-}
-
-/** Get the currently active isolation scope. */
-export function getIsolationScope(): Scope {
-  return getScopes().isolationScope as Scope;
-}
-
-/**
- * Set the currently active isolation scope.
- * Use this with caution! As it updates the isolation scope for the current execution context.
- */
-export function setIsolationScope(isolationScope: Scope): void {
-  getScopes().isolationScope = isolationScope;
-}
+export { getCurrentScope, getGlobalScope, getIsolationScope, getClient };
+export { setCurrentScope, setIsolationScope } from './scope';
 
 /**
  * Fork a scope from the current scope, and make it the current scope in the given callback
@@ -259,31 +209,6 @@ export function endSession(): void {
 
   // the session is over; take it off of the scope
   scope.setSession();
-}
-
-function getScopes(): CurrentScopes {
-  const carrier = getGlobalCarrier();
-
-  if (carrier.acs && carrier.acs.getScopes) {
-    const scopes = carrier.acs.getScopes();
-
-    if (scopes) {
-      return scopes;
-    }
-  }
-
-  return getGlobalCurrentScopes(carrier);
-}
-
-function getGlobalCurrentScopes(carrier: SentryCarrier): CurrentScopes {
-  if (!carrier.scopes) {
-    carrier.scopes = {
-      scope: new Scope(),
-      isolationScope: new Scope(),
-    };
-  }
-
-  return carrier.scopes;
 }
 
 /**
